requirements.txt
azure-data-tables
twilio
pyotp
structlog
pydantic
bs4
playwright
tzdata
